equipment-detector/requirements.txt
icrawler
tensorflow-macos==2.12.0
tensorflow-metal==0.8.0
numpy==1.23.5
tensorboard==2.12.3
keras==2.12.0
scipy
coremltools